capsuled game like login (see last commit)
now game is summed up like login, but because game now consists of only
one component (grid), it looks a little werid, but when other game ui
elements are implemented, it will start to make sense (like at the login
case)
