+++ app/components/AddPlayer.js_old
