--- app/containers/BackgroundContainer.js
@@ -1,18 +1,0 @@
-import { connect } from 'react-redux'
-
-import Background from '../components/Background'
-
-const mapStateToProps = (state) => {
-  return {}
-}
-
-const mapDispatchToProps = (dispatch) => {
-  return {}
-}
-
-const BackgroundContainer = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Background)
-
-export default BackgroundContainer
